Crate spin_factor_sqlite

Re-exports

pub use runtime_config::RuntimeConfig;

Modules

runtime_config

Structs

AppState

SqliteFactor

Constants

ALLOWED_DATABASES_KEY

Metadata key for a list of allowed databases for a component.

Traits

Connection

A trait abstracting over operations to a SQLite database

ConnectionCreator

A creator of a connections for a particular SQLite database.